refactor: extract page components from routes to feature directories
