+++ src/features/tools/pages/ThaiIdPage.tsx
@@ -1,0 +1,5 @@
+import ThaiIdGenerate from "@/features/tools/components/ThaiIdGenerate";
+
+export function ThaiIdPage() {
+  return <ThaiIdGenerate />;
+}
