+++ src/features/tools/pages/index.ts
@@ -1,1 +1,2 @@
 export * from "./ThaiNamesGeneratorPage";
+export * from "./ThaiIdPage";
