+++ src/routes/thai-id.tsx
@@ -1,9 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import ThaiIdGenerate from "@/features/tools/components/ThaiIdGenerate";
-
-function ThaiIdPage() {
-  return <ThaiIdGenerate />;
-}
+import { ThaiIdPage } from "@/features/tools/pages";
 
 export const Route = createFileRoute("/thai-id")({
   component: ThaiIdPage,
